+++ package.json
@@ -21,6 +21,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "4.18.2",
+    "jsonwebtoken": "^9.0.2",
     "morgan": "1.10.0",
     "zod": "^3.22.4"
   },
@@ -28,6 +29,7 @@
   "devDependencies": {
     "@types/cors": "^2.8.14",
     "@types/express": "^4.17.17",
+    "@types/jsonwebtoken": "^9.0.4",
     "@types/node": "^20.6.2",
     "@typescript-eslint/eslint-plugin": "^6.7.2",
     "@typescript-eslint/parser": "^6.7.2",
